[fix] Do parallel sums with int instead of bool.

+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -1924,7 +1924,7 @@ namespace Opm {
             }
         }
 
-        changed_well_to_group = comm.sum(changed_well_to_group);
+        changed_well_to_group = comm.sum(static_cast<int>(changed_well_to_group));
         if (changed_well_to_group) {
             updateAndCommunicate(episodeIdx, iterationIdx, deferred_logger);
             changed_well_group = true;
@@ -1939,7 +1939,7 @@ namespace Opm {
                 changed_well_individual = changed_well || changed_well_individual;
             }
         }
-        changed_well_individual = comm.sum(changed_well_individual);
+        changed_well_individual = comm.sum(static_cast<int>(changed_well_individual));
         if (changed_well_individual) {
             updateAndCommunicate(episodeIdx, iterationIdx, deferred_logger);
             changed_well_group = true;
